# additional/backend_2/static/swagger.yaml
openapi: 3.0.0
info:
  title: Diagnosis Decision Support System API
  version: "1.0"
paths:
  /Authentication/login:
    post:
      summary: User Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login Successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  role:
                    type: string
        '400':
          description: Error while authentication

  /Authentication/logout:
    post:
      summary: User Logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Log out Successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '400':
          description: Error while logging out
        '401':
          description: Unauthorized or JWT expired

  /EHR/overview:
    get:
      summary: Receptionist Overview
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Overview retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  patient_overview:
                    type: object
                  appointment_overview:
                    type: object
                  appointment_list:
                    type: array
                    items:
                      type: object
                  appointment_behavior:
                    type: array
                    items:
                      type: object
        '400':
          description: Error while retrieving overview
        '401':
          description: Unauthorized or JWT expired

  /EHR/patient_search:
    post:
      summary: Patient Search
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ssn:
                  type: string
                insurance_number:
                  type: string
      responses:
        '200':
          description: Patient search successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  patients:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                        ssn:
                          type: string
                        insurance_number:
                          type: string
        '400':
          description: Error while searching for patient
        '401':
          description: Unauthorized or JWT expired

  /EHR/retrieve:
    post:
      summary: EHR Retrieving
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                patient_id:
                  type: integer
      responses:
        '200':
          description: EHR retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  patient_information:
                    type: array
                    items:
                      type: object
                  pass_conditions:
                    type: array
                    items:
                      type: object
                  pass_sessions:
                    type: array
                    items:
                      type: object
        '400':
          description: Error while retrieving EHR
        '401':
          description: Unauthorized or JWT expired

  /EHR/check_in:
    post:
      summary: Check in for patient
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                patient_id:
                  type: integer
                reason_to_come:
                  type: string
                department:
                  type: string
                doctor:
                  type: string
      responses:
        '200':
          description: Check in Successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Error while checking in
        '401':
          description: Unauthorized or JWT expired

  /EHR/pass_session:
    post:
      summary: Pass Session Retrieving Module
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: integer
      responses:
        '200':
          description: Pass session retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  vital_sign:
                    type: array
                    items:
                      type: object
                  patient_information:
                    type: array
                    items:
                      type: object
                  test_results:
                    type: array
                    items:
                      type: object
                  symptoms:
                    type: string
                  diagnosis:
                    type: string
                  prescriptions:
                    type: array
                    items:
                      type: object
                  treatments:
                    type: array
                    items:
                      type: object
        '400':
          description: Error while retrieving pass session
        '401':
          description: Unauthorized or JWT expired

  /paraclinical/overview:
    get:
      summary: Paraclinical Overview
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Paraclinical overview retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  patient_overview:
                    type: object
                  patient_list:
                    type: array
                    items:
                      type: object
                  appointment_list:
                    type: array
                    items:
                      type: object
        '400':
          description: Error while retrieving paraclinical overview
        '401':
          description: Unauthorized or JWT expired

  /paraclinical/test:
    post:
      summary: Test View API
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                test_id:
                  type: integer
      responses:
        '200':
          description: Test view retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  pass_conditions:
                    type: array
                    items:
                      type: object
                  patient_information:
                    type: array
                    items:
                      type: object
                  pass_sessions:
                    type: array
                    items:
                      type: object
                  test_info:
                    type: array
                    items:
                      type: object
        '400':
          description: Error while retrieving test view
        '401':
          description: Unauthorized or JWT expired

  /paraclinical/test/detailed:
    post:
      summary: Detailed Test Result API
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                test_id:
                  type: integer
      responses:
        '200':
          description: Detailed test result retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  pass_conditions:
                    type: array
                    items:
                      type: object
                  patient_information:
                    type: array
                    items:
                      type: object
                  pass_sessions:
                    type: array
                    items:
                      type: object
                  test_info:
                    type: array
                    items:
                      type: object
                  test_result:
                    type: string
        '400':
          description: Error while retrieving detailed test result
        '401':
          description: Unauthorized or JWT expired

  /diagnosis/overview:
    get:
      summary: Doctor Overview
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Doctor overview retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  incoming_patient_list:
                    type: array
                    items:
                      type: object
                  paraclinical_patient_list:
                    type: array
                    items:
                      type: object
                  inpatient_list:
                    type: array
                    items:
                      type: object
                  appointment_list:
                    type: array
                    items:
                      type: object
        '400':
          description: Error while retrieving doctor overview
        '401':
          description: Unauthorized or JWT expired

  /diagnosis/recommend:
    post:
      summary: Recommend Prognosis API
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                symptoms:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: integer
                      name:
                        type: string
      responses:
        '200':
          description: Prognosis recommendation retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  recommendation:
                    type: array
                    items:
                      type: object
                      properties:
                        symptom:
                          type: string
                        probability:
                          type: number
        '400':
          description: Error while retrieving prognosis recommendation
        '401':
          description: Unauthorized or JWT expired

  /diagnosis/send_for_test:
    post:
      summary: Send Patient For Test API
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                test_type_id:
                  type: integer
                room_id:
                  type: integer
                note:
                  type: string
      responses:
        '200':
          description: Patient sent for test successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  recommendation:
                    type: array
                    items:
                      type: object
        '400':
          description: Error while sending patient for test
        '401':
          description: Unauthorized or JWT expired

  /diagnosis/finalizing:
    post:
      summary: Finalizing Diagnosis API
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                symptoms:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: integer
                      name:
                        type: string
                final_diagnosis:
                  type: string
      responses:
        '200':
          description: Diagnosis finalized successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Error while finalizing diagnosis
        '401':
          description: Unauthorized or JWT expired

  /diagnosis/prescription_treatment:
    post:
      summary: Prescription and Treatment API
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                prescriptions:
                  type: array
                  items:
                    type: object
                treatments:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: Prescription and treatment processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Error while processing prescription and treatment
        '401':
          description: Unauthorized or JWT expired

  /user_management/overview:
    get:
      summary: Admin Overview API
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Admin overview retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_list:
                    type: array
                    items:
                      type: object
                  active_users:
                    type: integer
                  inactive_users:
                    type: integer
        '400':
          description: Error while retrieving admin overview
        '401':
          description: Unauthorized or JWT expired

  /user_management/edit:
    post:
      summary: User Edit API
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                name:
                  type: string
                role:
                  type: string
      responses:
        '200':
          description: User edited successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Error while editing user
        '401':
          description: Unauthorized or JWT expired

  /user_management/add:
    post:
      summary: Add New User API
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                name:
                  type: string
                role:
                  type: string
      responses:
        '200':
          description: User added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Error while adding user
        '401':
          description: Unauthorized or JWT expired

  /user_management/delete:
    post:
      summary: Delete User API
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Error while deleting user
        '401':
          description: Unauthorized or JWT expired

  /user_management/user_logs:
    get:
      summary: View User Logs API
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User logs retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_logs:
                    type: array
                    items:
                      type: object
        '400':
          description: Error while retrieving user logs
        '401':
          description: Unauthorized or JWT expired

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT